+++ src/style/signup/signup.scss
@@ -5,10 +5,6 @@
   z-index: 1;
 }
 
-.plan-card {
-
-}
-
 .btn-large {
     width: 100%;
     height: 50px;
